+++ firebase-firestore/src/androidTest/java/com/google/firebase/firestore/FirestoreTest.java
@@ -971,10 +971,11 @@ public class FirestoreTest {
   }
 
   @Test
-  public void testClientCallsAfterShutdownFails() {
+  public void testClientCallsAfterTerminateFails() {
     FirebaseFirestore firestore = testFirestore();
-    waitFor(firestore.shutdown());
-    expectError(() -> waitFor(firestore.disableNetwork()), "The client has already been shutdown");
+    waitFor(firestore.terminate());
+    expectError(
+        () -> waitFor(firestore.disableNetwork()), "The client has already been terminated");
   }
 
   @Test
@@ -983,7 +984,7 @@ public class FirestoreTest {
         testFirestore(provider().projectId(), Level.DEBUG, newTestSettings(), "dbPersistenceKey");
     DocumentReference docRef = firestore.collection("col1").document("doc1");
     waitFor(docRef.set(map("foo", "bar")));
-    waitFor(AccessHelper.shutdown(firestore));
+    waitFor(firestore.terminate());
     IntegrationTestUtil.removeFirestore(firestore);
 
     // We restart the app with the same name and options to check that the previous instance's
@@ -1002,7 +1003,7 @@ public class FirestoreTest {
         testFirestore(provider().projectId(), Level.DEBUG, newTestSettings(), "dbPersistenceKey");
     DocumentReference docRef = firestore.collection("col1").document("doc1");
     waitFor(docRef.set(map("foo", "bar")));
-    waitFor(AccessHelper.shutdown(firestore));
+    waitFor(firestore.terminate());
     IntegrationTestUtil.removeFirestore(firestore);
     waitFor(AccessHelper.clearPersistence(firestore));
 
@@ -1038,7 +1039,7 @@ public class FirestoreTest {
     assertSame(instance, sameInstance);
     waitFor(instance.document("abc/123").set(Collections.singletonMap("field", 100L)));
 
-    instance.shutdown();
+    instance.terminate();
     FirebaseFirestore newInstance = FirebaseFirestore.getInstance(app);
 
     // Verify new instance works.
@@ -1051,25 +1052,25 @@ public class FirestoreTest {
   }
 
   @Test
-  public void testAppDeleteLeadsToFirestoreShutdown() {
+  public void testAppDeleteLeadsToFirestoreTerminate() {
     FirebaseApp app = testFirebaseApp();
     FirebaseFirestore instance = FirebaseFirestore.getInstance(app);
     waitFor(instance.document("abc/123").set(Collections.singletonMap("Field", 100)));
 
     app.delete();
 
-    assertTrue(instance.getClient().isShutdown());
+    assertTrue(instance.getClient().isTerminated());
   }
 
   @Test
-  public void testNewOperationThrowsAfterFirestoreShutdown() {
+  public void testNewOperationThrowsAfterFirestoreTerminate() {
     FirebaseFirestore instance = testFirestore();
     DocumentReference reference = instance.document("abc/123");
     waitFor(reference.set(Collections.singletonMap("Field", 100)));
 
-    instance.shutdown();
+    instance.terminate();
 
-    final String expectedMessage = "The client has already been shutdown";
+    final String expectedMessage = "The client has already been terminated";
     expectError(() -> waitFor(reference.get()), expectedMessage);
     expectError(() -> waitFor(reference.update("Field", 1)), expectedMessage);
     expectError(
@@ -1083,31 +1084,31 @@ public class FirestoreTest {
   }
 
   @Test
-  public void testShutdownCalledMultipleTimes() {
+  public void testTerminateCalledMultipleTimes() {
     FirebaseFirestore instance = testFirestore();
     DocumentReference reference = instance.document("abc/123");
     waitFor(reference.set(Collections.singletonMap("Field", 100)));
 
-    instance.shutdown();
+    instance.terminate();
 
-    final String expectedMessage = "The client has already been shutdown";
+    final String expectedMessage = "The client has already been terminated";
     expectError(() -> waitFor(reference.get()), expectedMessage);
 
     // Calling a second time should go through and change nothing.
-    instance.shutdown();
+    instance.terminate();
 
     expectError(() -> waitFor(reference.get()), expectedMessage);
   }
 
   @Test
-  public void testCanStopListeningAfterShutdown() {
+  public void testCanStopListeningAfterTerminate() {
     FirebaseFirestore instance = testFirestore();
     DocumentReference reference = instance.document("abc/123");
     EventAccumulator<DocumentSnapshot> eventAccumulator = new EventAccumulator<>();
     ListenerRegistration registration = reference.addSnapshotListener(eventAccumulator.listener());
     eventAccumulator.await();
 
-    waitFor(instance.shutdown());
+    waitFor(instance.terminate());
 
     // This should proceed without error.
     registration.remove();